refactor: split code into module.
